feat: add HistorialViewModal component for displaying medical history details
- Implemented a new modal component to show detailed medical history information.
- Added formatting for dates and rendering of file links.
- Enhanced the API client to ensure proper data structure handling in getFichaMedicaActual function.
